requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.0
groq==0.10.0
httpx==0.25.2
python-dotenv==1.0.1